fix sideEffect useInputs called before computed
This bug created a side effect that cause the order of some imports
to matter.  This surfaced after refactoring the vite plugin to clear
the mapStateUsed state array.  (Which caused other side effects)
After this was done, the controlled and uncontrolled forms tests
started failing, this was due to the fact that now useInputs was
being called before addAttributeHandler(attributeStateHandler);
This cause the order of attribute state handlers to be switched
(which is totally fine) however, the attributeStateHandler would
then take over for the input value state handler because it did
not properly check to see if useInputs had been called and if the
attribute was a "value" or "checked" attribute.